DOM seperation
