+++ src/DOM/projectDOM.js
@@ -1,8 +1,6 @@
-/* eslint-disable import/prefer-default-export */
-
 import { projectList } from '../modules/constructor';
 
-export const addProjectToTaskDropdown = (name, id) => {
+export default function addProjectToTaskDropdown(name, id) {
   const projectEntry = document.getElementById('projectEntry');
   const listProject = document.getElementById('listProject');
   const newOption1 = document.createElement('option');
@@ -13,4 +11,4 @@ export const addProjectToTaskDropdown = (name, id) => {
   newOption2.value = id;
   projectEntry.add(newOption1, projectList[id]);
   listProject.add(newOption2, listProject[id]);
-};
+}
